requirements.txt
# 基于 ResNet 的工业零件表面缺陷分类系统 - Python 依赖包

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 数据可视化
matplotlib>=3.5.0
numpy>=1.21.0

# 机器学习
scikit-learn>=1.0.0

# Web 应用框架
streamlit>=1.20.0

# 图像处理
Pillow>=9.0.0

# 进度条显示
tqdm>=4.60.0
